refactor: extract view specific navigation
